[DEBUG](Android) Fix float conversion issue in progress callback

+++ JNI/c/ARUPDATER_JNI_Downloader.c
@@ -286,7 +286,7 @@ void ARUPDATER_JNI_Downloader_ProgressCallback(void* arg, float percent)
         if ((ARUPDATER_JNI_Manager_VM != NULL) && (callbacks->jProgressListener != NULL) && (methodId_DownloaderListener_onPlfDownloadProgress != NULL))
         {
             JNIEnv *env = NULL;
-            jint jPercent = 0;
+            jfloat jPercent = 0;
             jint jResultEnv = 0;
             int error = JNI_OK;
